requirements.txt

python-dotenv
fastapi
uvicorn
openai